Getting query parameters should now be gc-safe - First we collect object values for each parameter, then we change it into a C string. The C strings are stored into internal pointers, so we store the objects into a CFRetained NSArray while processing them. - Also made -[PGTSQuery parameterCount] return an NSUInteger. - Fixes #194.
git-svn-id: http://basetenframework.org/svn/trunk@1197 fe1aa653-7022-0410-af67-bafd91c7e665

+++ Sources/PGTSQuery.h
@@ -55,7 +55,7 @@
 }
 - (NSArray *) parameters;
 - (void) setParameters: (NSArray *) anArray;
-- (int) parameterCount;
+- (NSUInteger) parameterCount;
 @end
